feat: use workers to reduce-literals
