+++ lib/reduce-literals.js
@@ -3,26 +3,7 @@
 const path = require('path');
 const CachingWriter = require('broccoli-caching-writer');
 const debug = require('debug');
-const fs = require('fs-extra');
-const { Transform } = require('stream');
-const mergeStream = require('merge-stream');
-const { chain } = require('stream-chain');
-const { parser } = require('stream-json');
-const { streamValues } = require('stream-json/streamers/StreamValues');
-
-const joinStream = () => new Transform({
-	transform(chunk, encoding, callback) {
-		this.data = this.data || [];
-		this.data.push(chunk.toString('utf8'));
-		callback();
-	},
-	flush() {
-		const value = JSON.stringify([...new Set(this.data)]);
-
-		this.push(value);
-		this.push(null);
-	}
-});
+const workerpool = require('workerpool');
 
 function groupByModule(entries, modules) {
 	return entries.reduce((acc, entry) => {
@@ -56,55 +37,20 @@ module.exports = class ReduceLiterals extends CachingWriter {
 		this.modules = options.modules;
 	}
 
-	async readEntries() {
-		this.debug('reading files...');
-
-		return this.listEntries();
-	}
-
-	async writeModule([module, paths]) {
-		const outputDir = path.join(this.outputPath, module);
-		const filename = path.join(outputDir, this.outputFile);
-
-		await fs.ensureDir(outputDir);
-
-		return new Promise((resolve, reject) => {
-			const pipelines = paths.map((entryPath) => chain([
-				fs.createReadStream(entryPath),
-				parser(),
-				// Parse each array entry
-				streamValues(),
-				// Trim strings
-				({ value }) => value.map((s) => s.trim()),
-				// Filter valid values
-				(value) => {
-					const isValid = value && value.match(/^[\w-]+(?:\.[\w-]+)+$/);
-
-					return isValid && value;
-				}
-			]));
-
-			const writeStream = chain([
-				// Merge all arrays in a single one
-				mergeStream(pipelines),
-				// Transform all entries in a stringified Array
-				joinStream(),
-				fs.createWriteStream(filename)
-			]);
-
-			writeStream.on('end', resolve);
-			writeStream.on('error', reject);
-		});
-	}
-
 	async build() {
 		const entries = this.listEntries();
 		const modules = groupByModule(entries, this.modules);
+		const pool = workerpool.pool(
+			path.join(__dirname, 'workers/reduce-literals.js')
+		);
 
 		this.debug('writing modules...');
 
-		for (const entry of Object.entries(modules)) {
-			await this.writeModule(entry);
-		}
+		const tasks = Object.entries(modules).map(([module, paths]) =>
+			pool.exec('writeModule', [this.outputPath, this.outputFile, module, paths])
+		);
+
+		await Promise.all(tasks);
+		pool.terminate();
 	}
 };
