+++ lib/workers/reduce-literals.js
@@ -1,27 +1,61 @@
 const workerpool = require('workerpool');
 const path = require('path');
-const readF = require('../utils/read-file');
-const writeF = require('../utils/write-file');
-const shouldWriteF = require('../utils/should-write-file');
+const fs = require('fs-extra');
+const { Transform } = require('stream');
+const mergeStream = require('merge-stream');
+const { chain } = require('stream-chain');
+const { parser } = require('stream-json');
+const { streamValues } = require('stream-json/streamers/StreamValues');
 
-async function readFile(entry) {
-	const filePath = path.join(entry.basePath, entry.relativePath);
-	const content = await readF(filePath);
-	const literals = JSON.parse(content);
+const joinStream = () => new Transform({
+	transform(chunk, encoding, callback) {
+		this.data = this.data || [];
+		this.data.push(chunk.toString('utf8'));
+		callback();
+	},
+	flush() {
+		const value = JSON.stringify([...new Set(this.data)]);
 
-	return [literals, entry.relativePath];
-}
-
-async function writeModule([module, literals]) {
-	const filename = path.join(this.outputPath, module, this.outputFile);
-	const content = JSON.stringify(literals);
-
-	if (await shouldWriteF(filename, content)) {
-		await writeF(filename, content);
+		this.push(value);
+		this.push(null);
 	}
+});
+
+async function writeModule(outputPath, outputFile, module, paths) {
+	const outputDir = path.join(outputPath, module);
+	const filename = path.join(outputDir, outputFile);
+
+	await fs.ensureDir(outputDir);
+
+	return new Promise((resolve, reject) => {
+		const pipelines = paths.map((entryPath) => chain([
+			fs.createReadStream(entryPath),
+			parser(),
+			// Parse each array entry
+			streamValues(),
+			// Trim strings
+			({ value }) => value.map((s) => s.trim()),
+			// Filter valid values
+			(value) => {
+				const isValid = value && value.match(/^[\w-]+(?:\.[\w-]+)+$/);
+
+				return isValid && value;
+			}
+		]));
+
+		const writeStream = chain([
+			// Merge all arrays in a single one
+			mergeStream(pipelines),
+			// Transform all entries in a stringified Array
+			joinStream(),
+			fs.createWriteStream(filename)
+		]);
+
+		writeStream.on('end', resolve);
+		writeStream.on('error', reject);
+	});
 }
 
 workerpool.worker({
-	readFile,
 	writeModule,
 });
